+++ sdk/avs/arm-avs/src/models/operationsMappers.ts
@@ -7,8 +7,9 @@
  */
 
 export {
-  ApiError,
-  ApiErrorBase,
+  CloudError,
+  ErrorAdditionalInfo,
+  ErrorResponse,
   Operation,
   OperationDisplay,
   OperationList