+++ sdk/keyvault/keyvault-keys/src/index.ts
@@ -52,13 +52,11 @@ import {
   CryptographyOptions,
   DeletedKey,
   DeleteKeyOptions,
-  EncryptionAlgorithm,
   GetDeletedKeyOptions,
   GetKeyOptions,
   ImportKeyOptions,
   JsonWebKey,
   KeyClientInterface,
-  KeyCurveName,
   KeyOperation,
   KeyPollerOptions,
   KeyType,
@@ -81,26 +79,30 @@ import {
 import {
   CryptographyClient,
   DecryptOptions,
-  DecryptResult,
   EncryptOptions,
-  EncryptResult,
-  SignatureAlgorithm,
-  KeyWrapAlgorithm,
   SignOptions,
-  SignResult,
   UnwrapKeyOptions,
-  UnwrapResult,
   VerifyOptions,
-  VerifyResult,
-  WrapKeyOptions,
-  WrapResult
+  WrapKeyOptions
 } from "./cryptographyClient";
 
+import { LocalCryptographyClient } from "./localCryptographyClient";
+
 import {
-  parseKeyVaultKeysIdentifier,
-  ParsedKeyVaultKeysIdentifier,
-  KeyVaultKeysIdentifierCollectionName
-} from "./identifier";
+  DecryptResult,
+  KeyCurveName,
+  EncryptionAlgorithm,
+  SignatureAlgorithm,
+  KeyWrapAlgorithm,
+  SignResult,
+  UnwrapResult,
+  VerifyResult,
+  WrapResult,
+  EncryptResult
+} from "./cryptographyClientModels";
+import { LocalSupportedAlgorithmName } from "./localCryptography/algorithms";
+
+import { parseKeyVaultKeyId, KeyVaultKeyId } from "./identifier";
 
 export {
   CryptographyClientOptions,
@@ -126,7 +128,7 @@ export {
   KeyOperation,
   KeyType,
   KeyPollerOptions,
-  parseKeyVaultKeysIdentifier,
+  parseKeyVaultKeyId,
   BeginDeleteKeyOptions,
   BeginRecoverDeletedKeyOptions,
   KeyProperties,
@@ -136,10 +138,11 @@ export {
   ListPropertiesOfKeysOptions,
   ListPropertiesOfKeyVersionsOptions,
   ListDeletedKeysOptions,
+  LocalCryptographyClient,
+  LocalSupportedAlgorithmName,
   PageSettings,
   PagedAsyncIterableIterator,
-  ParsedKeyVaultKeysIdentifier,
-  KeyVaultKeysIdentifierCollectionName,
+  KeyVaultKeyId,
   PipelineOptions,
   PollOperationState,
   PollerLike,
@@ -173,7 +176,7 @@ export class KeyClient {
   /**
    * @internal
    * @ignore
-   * A reference to the auto-generated KeyVault HTTP client.
+   * A reference to the auto-generated Key Vault HTTP client.
    */
   private readonly client: KeyVaultClient;
 
@@ -256,9 +259,9 @@ export class KeyClient {
   /**
    * @internal
    * @ignore
-   * Sends a delete request for the given KeyVault Key's name to the KeyVault service.
-   * Since the KeyVault Key won't be immediately deleted, we have {@link beginDeleteKey}.
-   * @param {string} name The name of the KeyVault Key.
+   * Sends a delete request for the given Key Vault Key's name to the Key Vault service.
+   * Since the Key Vault Key won't be immediately deleted, we have {@link beginDeleteKey}.
+   * @param {string} name The name of the Key Vault Key.
    * @param {DeleteKeyOptions} [options] Optional parameters for the underlying HTTP request.
    */
   private async deleteKey(name: string, options: DeleteKeyOptions = {}): Promise<DeletedKey> {
@@ -282,9 +285,9 @@ export class KeyClient {
   /**
    * @internal
    * @ignore
-   * Sends a request to recover a deleted KeyVault Key based on the given name.
-   * Since the KeyVault Key won't be immediately recover the deleted key, we have {@link beginRecoverDeletedKey}.
-   * @param {string} name The name of the KeyVault Key.
+   * Sends a request to recover a deleted Key Vault Key based on the given name.
+   * Since the Key Vault Key won't be immediately recover the deleted key, we have {@link beginRecoverDeletedKey}.
+   * @param {string} name The name of the Key Vault Key.
    * @param {RecoverDeletedKeyOptions} [options] Optional parameters for the underlying HTTP request.
    */
   private async recoverDeletedKey(
@@ -840,7 +843,7 @@ export class KeyClient {
    * @internal
    * @ignore
    * Deals with the pagination of {@link listPropertiesOfKeyVersions}.
-   * @param {string} name The name of the KeyVault Key.
+   * @param {string} name The name of the Key Vault Key.
    * @param {PageSettings} continuationState An object that indicates the position of the paginated request.
    * @param {ListPropertiesOfKeyVersionsOptions} [options] Common options for the iterative endpoints.
    */
@@ -883,7 +886,7 @@ export class KeyClient {
    * @internal
    * @ignore
    * Deals with the iteration of all the available results of {@link listPropertiesOfKeyVersions}.
-   * @param {string} name The name of the KeyVault Key.
+   * @param {string} name The name of the Key Vault Key.
    * @param {ListPropertiesOfKeyVersionsOptions} [options] Common options for the iterative endpoints.
    */
   private async *listPropertiesOfKeyVersionsAll(
@@ -1136,7 +1139,7 @@ export class KeyClient {
     const keyBundle = bundle as KeyBundle;
     const deletedKeyBundle = bundle as DeletedKeyBundle;
 
-    const parsedId = parseKeyVaultKeysIdentifier(keyBundle.key!.kid!);
+    const parsedId = parseKeyVaultKeyId(keyBundle.key!.kid!);
 
     const attributes: any = keyBundle.attributes || {};
     delete keyBundle.attributes;
@@ -1185,7 +1188,7 @@ export class KeyClient {
    * Shapes the exposed {@link DeletedKey} based on a received KeyItem.
    */
   private getDeletedKeyFromKeyItem(keyItem: KeyItem): DeletedKey {
-    const parsedId = parseKeyVaultKeysIdentifier(keyItem.kid!);
+    const parsedId = parseKeyVaultKeyId(keyItem.kid!);
 
     const attributes = keyItem.attributes || {};
 
@@ -1228,7 +1231,7 @@ export class KeyClient {
    * Shapes the exposed {@link KeyProperties} based on a received KeyItem.
    */
   private getKeyPropertiesFromKeyItem(keyItem: KeyItem): KeyProperties {
-    const parsedId = parseKeyVaultKeysIdentifier(keyItem.kid!);
+    const parsedId = parseKeyVaultKeyId(keyItem.kid!);
 
     const attributes = keyItem.attributes || {};