+++ sdk/search/search-documents/recordings/node/searchclient_uploaddocuments/recording_upload_a_set_of_documents.js
@@ -1,6 +1,6 @@
 let nock = require('nock');
 
-module.exports.hash = "c440ab9f23e146ad8d22c9a0dbf5220c";
+module.exports.hash = "feb0ac099aa6552b8f29163f3478ff61";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
@@ -20,9 +20,9 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Vary',
   'Accept-Encoding',
   'request-id',
-  'ca627c69-7fd7-4e94-bd95-eb7ecceaf9e0',
+  'ea33e178-79ee-4806-90a6-dafbc98b9133',
   'elapsed-time',
-  '36',
+  '42',
   'OData-Version',
   '4.0',
   'Preference-Applied',
@@ -30,7 +30,7 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=15724800; includeSubDomains',
   'Date',
-  'Wed, 12 Aug 2020 05:42:31 GMT',
+  'Wed, 26 Aug 2020 01:08:38 GMT',
   'Content-Length',
   '191' ]);
 
@@ -50,9 +50,9 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Vary',
   'Accept-Encoding',
   'request-id',
-  '58a50c40-51f8-4784-ab5e-d14f2f34311e',
+  '6065ddbb-2c6d-45fa-8824-35aae7d5c3e9',
   'elapsed-time',
-  '5',
+  '3',
   'OData-Version',
   '4.0',
   'Preference-Applied',
@@ -60,6 +60,6 @@ nock('https://endpoint:443', {"encodedQueryParams":true})
   'Strict-Transport-Security',
   'max-age=15724800; includeSubDomains',
   'Date',
-  'Wed, 12 Aug 2020 05:42:33 GMT',
+  'Wed, 26 Aug 2020 01:08:39 GMT',
   'Content-Length',
   '127' ]);