+++ sdk/storage/storage-blob-changefeed/src/index.ts
@@ -1,2 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
 export * from "./BlobChangeFeedClient";
 export * from "./models/BlobChangeFeedEvent";