+++ sdk/storage/storage-blob/src/utils/BufferScheduler.browser.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+// This file is used as a shim of "BufferScheduler" for some browser bundlers
+// when trying to bundle "BufferScheduler"
+// "BufferScheduler" class is only available in Node.js runtime
+export class BufferScheduler {}