+++ sdk/tables/data-tables/samples/javascript/src/connectionString.js
@@ -13,17 +13,23 @@ const connectionString = process.env["ACCOUNT_CONNECTION_STRING"] || "";
 async function listTables() {
   const client = TableServiceClient.fromConnectionString(connectionString);
 
-  const tables = await client.listTables();
+  const tables = client.listTables();
 
-  console.log(tables.value);
+  for await (let table of tables) {
+    console.log(table);
+    // {TableName: "TestTable"}
+  }
 }
 
 async function listEntities() {
   const client = TableClient.fromConnectionString(connectionString, "test1");
 
-  const entities = await client.listEntities();
+  const entities = client.listEntities();
 
-  console.log(entities.value);
+  for await (const entity of entities) {
+    console.log(entity);
+    // {PartitionKey: "P1", RowKey: "R1", foo: "Bar"}
+  }
 }
 
 async function main() {