+++ sdk/tables/data-tables/src/utils/fromAccountConnectionString.ts
@@ -4,6 +4,7 @@
 import { TableServiceClientOptions } from "..";
 import { ConnectionString } from "./connectionString";
 import { TablesSharedKeyCredential } from "../TablesSharedKeyCredential";
+import { createPipelineFromOptions } from "@azure/core-http";
 
 /**
  * Gets client parameters from an Account Connection String
@@ -11,21 +12,18 @@ import { TablesSharedKeyCredential } from "../TablesSharedKeyCredential";
  * @param extractedCreds parsed connection string
  * @param options TablesServiceClient options
  */
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function fromAccountConnectionString(
   extractedCreds: ConnectionString,
-  options?: TableServiceClientOptions
+  options: TableServiceClientOptions = {}
 ) {
   const sharedKeyCredential = new TablesSharedKeyCredential(
     extractedCreds.accountName!,
     extractedCreds.accountKey
   );
-  const optionsWithCredentials: TableServiceClientOptions = {
-    ...options,
-    requestPolicyFactories: (defaultFactories) => [...defaultFactories, sharedKeyCredential]
-  };
 
   return {
     url: extractedCreds.url,
-    options: optionsWithCredentials
+    options: createPipelineFromOptions(options, sharedKeyCredential)
   };
 }