+++ sdk/textanalytics/ai-text-analytics/recordings/node/aad_textanalyticsclient_recognizepiientities/recording_service_errors_on_unsupported_language.js
@@ -8,7 +8,7 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   .post('/azure_tenant_id/oauth2/v2.0/token', "response_type=token&grant_type=client_credentials&client_id=azure_client_id&client_secret=azure_client_secret&scope=https%3A%2F%2Fcognitiveservices.azure.com%2F.default")
   .reply(200, {"token_type":"Bearer","expires_in":3599,"ext_expires_in":3599,"access_token":"access_token"}, [
   'Cache-Control',
-  'no-cache, no-store',
+  'no-store, no-cache',
   'Pragma',
   'no-cache',
   'Content-Type',
@@ -19,39 +19,40 @@ nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
   'max-age=31536000; includeSubDomains',
   'X-Content-Type-Options',
   'nosniff',
-  'x-ms-request-id',
-  'b07ee0aa-5391-4adf-b6fc-348da3de2700',
-  'x-ms-ests-server',
-  '2.1.10946.15 - SCUS ProdSlices',
   'P3P',
   'CP="DSP CUR OTPi IND OTRi ONL FIN"',
+  'x-ms-request-id',
+  '8ed8a733-5154-44dc-8031-7800d29f3800',
+  'x-ms-ests-server',
+  '2.1.10985.18 - CHI ProdSlices',
   'Set-Cookie',
-  'fpc=ApGhofnbuZNNug2ZR81RpyX0CyfMAQAAAJzFyNYOAAAA; expires=Sun, 13-Sep-2020 17:49:16 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'fpc=ApSkrxRsYpRPnuiIlMxO9FzIIHRUAQAAAJEn4tYOAAAA; expires=Fri, 02-Oct-2020 23:53:53 GMT; path=/; secure; HttpOnly; SameSite=None',
   'Set-Cookie',
-  'x-ms-gateway-slice=estsfd; path=/; SameSite=None; secure; HttpOnly',
+  'x-ms-gateway-slice=prod; path=/; secure; samesite=none; httponly',
   'Set-Cookie',
-  'stsservicecookie=estsfd; path=/; SameSite=None; secure; HttpOnly',
+  'stsservicecookie=ests; path=/; secure; samesite=none; httponly',
   'Date',
-  'Fri, 14 Aug 2020 17:49:16 GMT',
+  'Wed, 02 Sep 2020 23:53:53 GMT',
   'Content-Length',
-  '1500'
+  '1329'
 ]);
 
 nock('https://endpoint', {"encodedQueryParams":true})
   .post('/text/analytics/v3.1-preview.1/entities/recognition/pii', {"documents":[{"id":"0","text":"This is some text, but it doesn't matter.","language":"notalanguage"}]})
+  .query(true)
   .reply(200, {"documents":[],"errors":[{"id":"0","error":{"code":"InvalidArgument","message":"Invalid Language Code.","innererror":{"code":"UnsupportedLanguageCode","message":"Invalid language code. Supported languages: en"}}}],"modelVersion":"2020-07-01"}, [
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
   'application/json; charset=utf-8',
   'x-envoy-upstream-service-time',
-  '811',
+  '2',
   'apim-request-id',
-  'e9f329c0-5400-4946-8cc6-a36250f3dabe',
+  '596cc7b4-7268-4722-8d2c-cb0573f7054e',
   'Strict-Transport-Security',
   'max-age=31536000; includeSubDomains; preload',
   'x-content-type-options',
   'nosniff',
   'Date',
-  'Fri, 14 Aug 2020 17:49:17 GMT'
+  'Wed, 02 Sep 2020 23:53:52 GMT'
 ]);