+++ sdk/textanalytics/ai-text-analytics/recordings/node/aad_textanalyticsclient_string_encoding/recording_emoji.js
@@ -1,0 +1,60 @@
+let nock = require('nock');
+
+module.exports.hash = "32979b66725b7de4f48ed06ee2bf82fb";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
+  .post('/azure_tenant_id/oauth2/v2.0/token', "response_type=token&grant_type=client_credentials&client_id=azure_client_id&client_secret=azure_client_secret&scope=https%3A%2F%2Fcognitiveservices.azure.com%2F.default")
+  .reply(200, {"token_type":"Bearer","expires_in":3599,"ext_expires_in":3599,"access_token":"access_token"}, [
+  'Cache-Control',
+  'no-store, no-cache',
+  'Pragma',
+  'no-cache',
+  'Content-Length',
+  '1329',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Expires',
+  '-1',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains',
+  'X-Content-Type-Options',
+  'nosniff',
+  'P3P',
+  'CP="DSP CUR OTPi IND OTRi ONL FIN"',
+  'x-ms-request-id',
+  'f17ec1d1-63b7-4d36-9ef5-cd736f3b5200',
+  'x-ms-ests-server',
+  '2.1.10985.18 - CHI ProdSlices',
+  'Set-Cookie',
+  'fpc=Amx6cLU0rTdMlQzayoICOUbIIHRUAQAAAJQn4tYOAAAA; expires=Fri, 02-Oct-2020 23:53:57 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'Set-Cookie',
+  'x-ms-gateway-slice=prod; path=/; secure; samesite=none; httponly',
+  'Set-Cookie',
+  'stsservicecookie=ests; path=/; secure; samesite=none; httponly',
+  'Date',
+  'Wed, 02 Sep 2020 23:53:56 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/text/analytics/v3.1-preview.1/entities/recognition/pii', {"documents":[{"id":"0","text":"👩 SSN: 859-98-0987","language":"en"}]})
+  .query(true)
+  .reply(200, {"documents":[{"id":"0","entities":[{"text":"859-98-0987","category":"U.S. Social Security Number (SSN)","offset":8,"length":11,"confidenceScore":0.65}],"warnings":[]}],"errors":[],"modelVersion":"2020-07-01"}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'csp-billing-usage',
+  'CognitiveServices.TextAnalytics.BatchScoring=1',
+  'x-envoy-upstream-service-time',
+  '72',
+  'apim-request-id',
+  '27bb67f4-c37e-456d-8f66-6e2fd0fefc60',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains; preload',
+  'x-content-type-options',
+  'nosniff',
+  'Date',
+  'Wed, 02 Sep 2020 23:53:56 GMT'
+]);